requirements.txt
pymupdf==1.22.5
pytesseract==0.3.10
Pillow==9.4.0
streamlit==1.24.1
langchain==0.0.283
google-generativeai==0.2.1
python-dotenv==1.0.0
pdfplumber==0.9.0